refactor converter
